backend/requirements.txt
Flask
flask-socketio
flask-cors
python-dotenv
requests
boto3
langchain
langchain-google-genai
google-generativeai
faiss-cpu
numpy
eventlet
gunicorn
gevent
gevent-websocket
fastapi
uvicorn
aiofiles
requests
boto3
faiss-cpu
openai
langchain
python-dotenv
fastapi
uvicorn
langchain_community
Flask
flask-socketio
flask-cors
python-dotenv
requests
boto3
langchain
langchain-google-genai
google-generativeai
faiss-cpu
numpy
eventlet
gunicorn
gevent
gevent-websocket
fastapi
uvicorn
aiofiles
openai
langchain_community
python-multipart